+++ lib/main.dart
@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:schrodinger_client/account/accountbank.dart';
 
-import 'package:schrodinger_client/accountbank.dart';
-import 'package:schrodinger_client/budgetpage.dart';
-import 'package:schrodinger_client/expensepage.dart';
+import 'package:schrodinger_client/account/budgetpage.dart';
+import 'package:schrodinger_client/account/expense_category_page.dart';
+import 'package:schrodinger_client/account/expensepage.dart';
+import 'package:schrodinger_client/account/income_category.dart';
+import 'package:schrodinger_client/account/income_page.dart';
+import 'package:schrodinger_client/account/income_records.dart';
+import 'package:schrodinger_client/account/payment_methods.dart';
+import 'package:schrodinger_client/account/spending_records.dart';
 import 'package:schrodinger_client/home.dart';
 import 'package:schrodinger_client/login/join_page.dart';
-import 'package:schrodinger_client/payment_methods.dart';
-import 'package:schrodinger_client/spending_records.dart';
-import 'package:schrodinger_client/income_category.dart';
-import 'package:schrodinger_client/income_records.dart';
-import 'package:schrodinger_client/expense_category_page.dart';
-import 'package:schrodinger_client/income_page.dart';
 import 'package:schrodinger_client/login/auth_page.dart';
 import 'package:schrodinger_client/login/login_page.dart';
 import 'package:schrodinger_client/login/splash.dart';
